requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
requests==2.31.0
apscheduler==3.10.4
python-dotenv==1.0.0
httpx==0.25.2
scipy==1.11.4
statsmodels==0.14.0
plotly==5.18.0
websockets==12.0
pydantic==2.5.2
alembic==1.12.1
loguru==0.7.2
pytest==7.4.3
fredapi==0.5.2
openpyxl>=3.1.0
